airflow-requirements.txt
# Airflow Requirements
apache-airflow==2.7.0
psycopg2-binary>=2.9.0
redis>=4.0.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Crypto API clients
cryptocompare>=0.7.6
ccxt>=4.0.0

# Database connectors
sqlalchemy>=1.4.0
alembic>=1.8.0

# Monitoring and logging
prometheus-client>=0.15.0
structlog>=22.0.0

# Data validation
pydantic>=1.10.0
jsonschema>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.0
croniter>=1.3.0
